Fold root-package stub files into the VFS tools.
richtext_adapters.go was a 12-line init, and tools_command.go only defined
run_command. Both belong with read/write on tools_vfs.go. Hosts still mount
backends and WriteFile to those paths; the agent uses the same VFS tools.
